pyproject.toml
[project]
name = "automated-sudoku"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest-playwright (>=0.7.0,<0.8.0)",
    "playwright (>=1.50.0,<2.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "python-constraint (>=1.4.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false